requirements.txt
fastapi[all]
torch
torchvision
requests
Pillow
python-multipart
matplotlib
numpy
opencv-python